+++ sql/core/src/main/scala/org/apache/spark/sql/DataFrameReader.scala
@@ -192,7 +192,8 @@ class DataFrameReader private[sql](sparkSession: SparkSession) extends Logging {
         "read files of Hive data source directly.")
     }
 
-    val disabledV2Readers = sparkSession.sessionState.conf.disabledV2FileReads.split(",")
+    val disabledV2Readers =
+      sparkSession.sessionState.conf.disabledV2FileReads.toLowerCase(Locale.ROOT).split(",")
     val lookupCls = DataSource.lookupDataSource(source, sparkSession.sessionState.conf)
     val cls = lookupCls.newInstance() match {
       case f: FileDataSourceV2 if disabledV2Readers.contains(f.shortName()) => f.fallBackFileFormat
